requirements.txt
Django>=3.2,<=5.1
djangorestframework>=3.12.0
django-filter>=21.0
django-cors-headers>=3.10.0
redis>=4.0.0
celery>=5.0.0
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=3.0.0
ruff>=0.1.0